+++ packages/koni_jxl/lib/koni_jxl.dart
@@ -5,6 +5,7 @@ export 'src/decoder.dart' show JxlDecoder;
 export 'src/encode/encoder.dart' show JxlEncoder;
 export 'src/encode/vardct/vardct_l0_encoder.dart' show VardctL0Config;
 export 'src/exceptions.dart';
+export 'src/io/container.dart' show looksLikeJxl;
 export 'src/jxl_image.dart' show JxlAnimation, JxlImage;
 export 'src/jxl_info.dart' show JxlInfo;
 export 'src/jxl_limits.dart' show JxlLimits;